src/arnold-journal-mcp/pyproject.toml
[project]
name = "arnold-journal-mcp"
version = "0.1.0"
description = "Arnold Journal MCP - Subjective data capture for fitness coaching"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "neo4j>=5.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
arnold-journal-mcp = "arnold_journal_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
